requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
coverage==6.4.3
Django==3.0.1
django-background-tasks==1.2.5
django-compat==1.0.15
djongo==1.3.6
idna==3.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pymongo==3.12.1
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.12
